fix: share the screenshot capture pipeline across linux and windows, restoring the scaleImage min-size clamps, region validation, and capture-error context on Windows
